Compile source with Google Closure Compiler
This changeset consists of several parts.

First, in order to get the most out of GCC's advanced optimisations, I need to
make use of its ability to parse ES6 (`language_in=ECMASCRIPT6`). This means
that the es2015 Babel transformations are not only superfluous, but are
detrimental to GCC's optimiser. After trying Rollup and the
`transform-es2015-modules-commonjs` Babel transformer, I decided that was using
CommonJS was the path of least resistance.

Next, GCC's advanced optimisations rename object properties. This is especially
for this codebase not only because of the interoperation with other libraries,
but also because of the consumption of remote JSON files. To work around this,
much of the object property notation has been converted to use `obj['prop']`
rather than `obj.prop` notation.

Finally, Webpack's default minimize optimisation is disabled. Since `-p`
implies `--optimize-minimize --optimize-occurence-order`, I now run
`webpack --optimize-occurence-order` and allow the GCC Webpack plugin to
perform its own minimize optimisations.